+++ EasyAssertions/Assertions/CollectionAssertions.cs
@@ -33,7 +33,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence has no elements in it.
         /// </summary>
-        public static Actual<TActual> ShouldBeEmpty<TActual>(this TActual actual, string message = null) 
+        public static Actual<TActual> ShouldBeEmpty<TActual>(this TActual actual, string message = null)
             where TActual : IEnumerable
         {
             return actual.RegisterAssertion(c =>
@@ -51,7 +51,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence contains at least one element.
         /// </summary>
-        public static Actual<TActual> ShouldNotBeEmpty<TActual>(this TActual actual, string message = null) 
+        public static Actual<TActual> ShouldNotBeEmpty<TActual>(this TActual actual, string message = null)
             where TActual : IEnumerable
         {
             return actual.RegisterAssertion(c =>
@@ -65,16 +65,33 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence contains exactly one element.
         /// </summary>
-        public static Actual<TActual> ShouldBeSingular<TActual>(this TActual actual, string message = null) 
+        public static Actual<TActual> ShouldBeSingular<TActual>(this TActual actual, string message = null)
             where TActual : IEnumerable
         {
             return actual.RegisterAssertion(c => actual.ShouldBeLength(1, message));
         }
 
         /// <summary>
+        /// Asserts that a sequence contains exactly one element of the specified type.
+        /// </summary>
+        public static Actual<TExpected> ShouldBeASingular<TExpected>(this IEnumerable actual, string message = null)
+        {
+            return actual.RegisterAssertion(c =>
+                {
+                    actual.ShouldBeA<IEnumerable>(message);
+                    using (IBuffer<object> bufferedActual = actual.Buffer())
+                    {
+                        bufferedActual.ShouldBeSingular(message);
+                        ObjectAssertions.AssertType<TExpected>(bufferedActual.Single(), message, c);
+                        return new Actual<TExpected>((TExpected)bufferedActual.Single());
+                    }
+                });
+        }
+
+        /// <summary>
         /// Asserts that a sequence contains a specific number of elements.
         /// </summary>
-        public static Actual<TActual> ShouldBeLength<TActual>(this TActual actual, int expectedLength, string message = null) 
+        public static Actual<TActual> ShouldBeLength<TActual>(this TActual actual, int expectedLength, string message = null)
             where TActual : IEnumerable
         {
             return actual.RegisterAssertion(c =>
@@ -94,7 +111,7 @@ namespace EasyAssertions
         /// <see cref="IEnumerable"/> items are compared recursively.
         /// Non-<c>IEnumerable</c> items are compared using the default equality comparer.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ShouldMatch<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldMatch<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null)
             where TExpected : TActual
         {
             if (expected == null) throw new ArgumentNullException(nameof(expected));
@@ -139,7 +156,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence starts with another sub-sequence, using the default equality comparer.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ShouldStartWith<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expectedStart, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldStartWith<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expectedStart, string message = null)
             where TExpected : TActual
         {
             if (expectedStart == null) throw new ArgumentNullException(nameof(expectedStart));
@@ -157,7 +174,7 @@ namespace EasyAssertions
                 });
         }
 
-        public static Actual<IEnumerable<TActual>> ShouldEndWith<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expectedEnd, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldEndWith<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expectedEnd, string message = null)
             where TExpected : TActual
         {
             if (expectedEnd == null) throw new ArgumentNullException(nameof(expectedEnd));
@@ -178,7 +195,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence contains a specified element, using the default equality comparer.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ShouldContain<TActual, TExpected>(this IEnumerable<TActual> actual, TExpected expected, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldContain<TActual, TExpected>(this IEnumerable<TActual> actual, TExpected expected, string message = null)
             where TExpected : TActual
         {
             return actual.RegisterAssertion(c =>
@@ -196,7 +213,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence contains all specified elements, in any order, using the default equality comparer.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ShouldContainItems<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldContainItems<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null)
             where TExpected : TActual
         {
             return actual.RegisterAssertion(c =>
@@ -215,7 +232,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that all elements in a sequence are contained within another sequence, in any order, using the default equality comparer.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ItemsShouldBeIn<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expectedSuperset, string message = null) 
+        public static Actual<IEnumerable<TActual>> ItemsShouldBeIn<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expectedSuperset, string message = null)
             where TExpected : TActual
         {
             if (expectedSuperset == null) throw new ArgumentNullException(nameof(expectedSuperset));
@@ -236,7 +253,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence does not contain a specified element, using the default equality comparer.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ShouldNotContain<TActual, TExpected>(this IEnumerable<TActual> actual, TExpected expectedToNotContain, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldNotContain<TActual, TExpected>(this IEnumerable<TActual> actual, TExpected expectedToNotContain, string message = null)
             where TExpected : TActual
         {
             return actual.RegisterAssertion(c =>
@@ -254,7 +271,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence does not contain any of the specified elements, using the default equality comparer.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ShouldNotContainItems<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expectedToNotContain, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldNotContainItems<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expectedToNotContain, string message = null)
             where TExpected : TActual
         {
             return actual.RegisterAssertion(c =>
@@ -273,7 +290,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts thats a sequence only contains the specified elements, and nothing else, in any order, with no duplicates, using the default equality comparer.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ShouldOnlyContain<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldOnlyContain<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null)
             where TExpected : TActual
         {
             return actual.RegisterAssertion(c =>
@@ -294,7 +311,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence contains at least one item that does appear in another specified sequence, using the default equality comparer.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ShouldNotOnlyContain<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldNotOnlyContain<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null)
             where TExpected : TActual
         {
             if (expected == null) throw new ArgumentNullException(nameof(expected));
@@ -315,13 +332,13 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that a sequence contains no duplicated elements, using the default equality comparer.
         /// </summary>
-        public static Actual<TActual> ShouldBeDistinct<TActual>(this TActual actual, string message = null) 
+        public static Actual<TActual> ShouldBeDistinct<TActual>(this TActual actual, string message = null)
             where TActual : IEnumerable
         {
             return actual.RegisterAssertion(c => AssertDistinct(c, actual, message));
         }
 
-        private static void AssertDistinct<TActual>(IAssertionContext assertionContext, TActual actual, string message) 
+        private static void AssertDistinct<TActual>(IAssertionContext assertionContext, TActual actual, string message)
             where TActual : IEnumerable
         {
             actual.ShouldBeA<TActual>(message);
@@ -336,7 +353,7 @@ namespace EasyAssertions
         /// <summary>
         /// Asserts that two sequences contain the same object instances in the same order.
         /// </summary>
-        public static Actual<IEnumerable<TActual>> ShouldMatchReferences<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null) 
+        public static Actual<IEnumerable<TActual>> ShouldMatchReferences<TActual, TExpected>(this IEnumerable<TActual> actual, IEnumerable<TExpected> expected, string message = null)
             where TExpected : TActual
         {
             return actual.RegisterAssertion(c =>